backend/k8s-misconfiguration/requirements.txt
Flask==3.0.3
GitPython==3.1.43
